Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main(){
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
-    int d= (a+b+c)/3;
-    printf("Average: %d\n", d);
+    float d= (a+b+c)/3;
+    printf("Average: %.2f\n", d);
     return 0;
 }
